requirements.txt
aiohttp>=3.9.0
websockets>=12.0
python-dotenv>=1.0.0
py-clob-client>=0.14.0
